Add a new way to parse vtkCommand.h...
...since the list of events in there completely changed form and
could not be parsed using the same regular expression that used
to pick out enum value names. Allow for both forms so this same
ActiViz code can be used to build against older VTK's and also
this new style.

First, we try to parse lines like this:
  AnyEvent,
  DeleteEvent,
  ...

Then, if none match, we parse lines like this:
    _vtk_add_event(AnyEvent)\
    _vtk_add_event(DeleteEvent)\
    ...

Both styles of parsing are limited to sections between matching
"begin" and "end" lines as well.